% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{dispersionPlot}
\alias{dispersionPlot}
\title{Quality Control: Dispersion plot}
\usage{
dispersionPlot(cuff)
}
\arguments{
\item{cuff}{cummeRbund cuff object}
}
\value{
Plot object
}
\description{
Quality Control: Dispersion plot
}
