% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{fpkmSCVPlotWrap}
\alias{fpkmSCVPlotWrap}
\title{Cross replicate variability}
\usage{
fpkmSCVPlotWrap(cuff, type = "genes")
}
\arguments{
\item{cuff}{cummeRbund cuff object}

\item{type}{Type can be either "genes" or "isoforms"}
}
\value{
Plot object
}
\description{
Cross replicate variability
}
