% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{read_cuffdiff}
\alias{read_cuffdiff}
\title{Read in Cuffdiff result}
\usage{
read_cuffdiff(path)
}
\arguments{
\item{path}{Path to cuffdiff folder}
}
\value{
cummeRbund cuff object
}
\description{
Read in Cuffdiff result
}
