% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{saveAndPrint}
\alias{saveAndPrint}
\title{Save ggplot and print it}
\usage{
saveAndPrint(plot, save_plot = TRUE, plot_name = "", output_path = "")
}
\arguments{
\item{plot}{ggplot2 plot object}

\item{save_plot}{Save plot to disk}

\item{plot_name}{name of ggplot file}

\item{output_path}{path to output}
}
\description{
Save ggplot and print it
}
