% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{saveReturnPlot}
\alias{saveReturnPlot}
\title{Save image and print it}
\usage{
saveReturnPlot(plot, save_plot = TRUE, plot_name = "", output_path = "",
  is_ggplot = TRUE)
}
\arguments{
\item{plot}{ggplot2 plot object}

\item{save_plot}{Save plot to disk}

\item{plot_name}{name of ggplot file}

\item{output_path}{path to output}

\item{is_ggplot}{Plot is ggplot object or not}
}
\description{
Save image and print it
}
